pyproject.toml
[tool.poetry]
name = "hw-web-11"
version = "0.1.0"
description = ""
package-mode = false
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.29"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.1"
libgravatar = "^1.0.4"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.40.0"
sphinx = "^7.3.7"
pytest = "^8.2.0"
httpx = "^0.27.0"

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
